LapinRobot/requirements.txt
pyserial
matplotlib
numpy
time
pyqt5
scipy